requirements.txt
pydub==0.25.1
pyannote.audio==3.3.2
